ai-training/requirements.txt
ultralytics
opencv-python-headless
numpy<2.0.0
torch
torchvision
pyyaml
matplotlib
pandas
